#3074-refactored announcements to be on separate route
