+++ src/backend/src/services/announcement.service.ts
@@ -2,6 +2,7 @@ import { Announcement } from 'shared';
 import prisma from '../prisma/prisma';
 import { getAnnouncementQueryArgs } from '../prisma-query-args/announcements.query.args';
 import announcementTransformer from '../transformers/announcements.transformer';
+import { HttpException } from '../utils/errors.utils';
 
 export default class AnnouncementService {
   /**
@@ -40,5 +41,26 @@ export default class AnnouncementService {
     });
 
     return announcementTransformer(announcement);
+  }
+
+  /**
+   * Gets all of a user's unread announcements
+   * @param userId id of the current user
+   * @param organization the user's orgainzation
+   * @returns the unread announcements of the user
+   */
+  static async getUserUnreadAnnouncements(userId: string, organizationId: string) {
+    const unreadAnnouncements = await prisma.announcement.findMany({
+      where: {
+        usersReceived: {
+          some: { userId }
+        }
+      },
+      ...getAnnouncementQueryArgs(organizationId)
+    });
+
+    if (!unreadAnnouncements) throw new HttpException(404, 'User Unread Announcements Not Found');
+
+    return unreadAnnouncements.map(announcementTransformer);
   }
 }
